Move compatibility changes to internal, use when pdata parses traces from bytes
