+++ receiver/otlpreceiver/trace/otlp_test.go
@@ -179,118 +179,3 @@ func otlpReceiverOnGRPCServer(t *testing.T, tc consumer.Traces) (int, func()) {
 
 	return port, done
 }
-
-func TestDeprecatedStatusCode(t *testing.T) {
-	traceSink := new(consumertest.TracesSink)
-
-	port, doneFn := otlpReceiverOnGRPCServer(t, traceSink)
-	defer doneFn()
-
-	traceClient, traceClientDoneFn, err := makeTraceServiceClient(port)
-	require.NoError(t, err, "Failed to create the TraceServiceClient: %v", err)
-	defer traceClientDoneFn()
-
-	// See specification for handling status code here:
-	// https://github.com/open-telemetry/opentelemetry-proto/blob/59c488bfb8fb6d0458ad6425758b70259ff4a2bd/opentelemetry/proto/trace/v1/trace.proto#L231
-	tests := []struct {
-		sendCode               otlptrace.Status_StatusCode
-		sendDeprecatedCode     otlptrace.Status_DeprecatedStatusCode
-		expectedRcvCode        otlptrace.Status_StatusCode
-		expectedDeprecatedCode otlptrace.Status_DeprecatedStatusCode
-	}{
-		{
-			// If code==STATUS_CODE_UNSET then the value of `deprecated_code` is the
-			//   carrier of the overall status according to these rules:
-			//
-			//     if deprecated_code==DEPRECATED_STATUS_CODE_OK then the receiver MUST interpret
-			//     the overall status to be STATUS_CODE_UNSET.
-			sendCode:               otlptrace.Status_STATUS_CODE_UNSET,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_UNSET,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-		},
-		{
-			//     if deprecated_code!=DEPRECATED_STATUS_CODE_OK then the receiver MUST interpret
-			//     the overall status to be STATUS_CODE_ERROR.
-			sendCode:               otlptrace.Status_STATUS_CODE_UNSET,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_ABORTED,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_ERROR,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_ABORTED,
-		},
-		{
-			//   If code!=STATUS_CODE_UNSET then the value of `deprecated_code` MUST be
-			//   overwritten, the `code` field is the sole carrier of the status.
-			sendCode:               otlptrace.Status_STATUS_CODE_OK,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_OK,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-		},
-		{
-			//   If code!=STATUS_CODE_UNSET then the value of `deprecated_code` MUST be
-			//   overwritten, the `code` field is the sole carrier of the status.
-			sendCode:               otlptrace.Status_STATUS_CODE_OK,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_UNKNOWN_ERROR,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_OK,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-		},
-		{
-			//   If code!=STATUS_CODE_UNSET then the value of `deprecated_code` MUST be
-			//   overwritten, the `code` field is the sole carrier of the status.
-			sendCode:               otlptrace.Status_STATUS_CODE_ERROR,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_OK,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_ERROR,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_UNKNOWN_ERROR,
-		},
-		{
-			//   If code!=STATUS_CODE_UNSET then the value of `deprecated_code` MUST be
-			//   overwritten, the `code` field is the sole carrier of the status.
-			sendCode:               otlptrace.Status_STATUS_CODE_ERROR,
-			sendDeprecatedCode:     otlptrace.Status_DEPRECATED_STATUS_CODE_UNKNOWN_ERROR,
-			expectedRcvCode:        otlptrace.Status_STATUS_CODE_ERROR,
-			expectedDeprecatedCode: otlptrace.Status_DEPRECATED_STATUS_CODE_UNKNOWN_ERROR,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.sendCode.String()+"/"+test.sendDeprecatedCode.String(), func(t *testing.T) {
-			resourceSpans := []*otlptrace.ResourceSpans{
-				{
-					InstrumentationLibrarySpans: []*otlptrace.InstrumentationLibrarySpans{
-						{
-							Spans: []*otlptrace.Span{
-								{
-									Status: otlptrace.Status{
-										Code:           test.sendCode,
-										DeprecatedCode: test.sendDeprecatedCode,
-									},
-								},
-							},
-						},
-					},
-				},
-			}
-
-			req := &collectortrace.ExportTraceServiceRequest{
-				ResourceSpans: resourceSpans,
-			}
-
-			traceSink.Reset()
-
-			resp, err := traceClient.Export(context.Background(), req)
-			require.NoError(t, err, "Failed to export trace: %v", err)
-			require.NotNil(t, resp, "The response is missing")
-
-			require.Equal(t, 1, len(traceSink.AllTraces()), "unexpected length: %v", len(traceSink.AllTraces()))
-
-			rcvdStatus := traceSink.AllTraces()[0].ResourceSpans().At(0).InstrumentationLibrarySpans().At(0).Spans().At(0).Status()
-
-			// Check that Code is as expected.
-			assert.EqualValues(t, rcvdStatus.Code(), test.expectedRcvCode)
-
-			spanProto := internal.TracesToOtlp(traceSink.AllTraces()[0].InternalRep()).ResourceSpans[0].InstrumentationLibrarySpans[0].Spans[0]
-
-			// Check that DeprecatedCode is passed as is.
-			assert.EqualValues(t, spanProto.Status.DeprecatedCode, test.expectedDeprecatedCode)
-		})
-	}
-}
